+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-03-18
+
 ### Added
 
 * The name of the PVC to be checked out is now configurable.
@@ -50,7 +52,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 * Initial release.
 
-[Unreleased]: https://github.com/puppetlabs/pvpool/compare/v0.1.4...HEAD
+[Unreleased]: https://github.com/puppetlabs/pvpool/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/puppetlabs/pvpool/compare/v0.1.4...v0.2.0
 [0.1.4]: https://github.com/puppetlabs/pvpool/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/puppetlabs/pvpool/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/puppetlabs/pvpool/compare/v0.1.1...v0.1.2